% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_clt.R
\name{demo_clt}
\alias{demo_clt}
\title{Demonstrate the Central Limit Theorem}
\usage{
demo_clt(rng, n, nrep = 10000, ..., mean = NULL, sd = NULL)
}
\arguments{
\item{rng}{A random number generator function (e.g., `runif`,
`rnorm`, `rgamma`).}

\item{n}{A numeric vector of sample sizes (e.g., `c(5, 10,
20, 40)`).}

\item{nrep}{The number of repetitions for generating sample
means (default is 10000).}

\item{...}{Additional arguments passed to the random number
generator (e.g., `shape` and `rate` for `rgamma`).}

\item{mean}{The theoretical mean of the distribution. If not
provided, it is estimated from a large sample.}

\item{sd}{The theoretical standard deviation of the
distribution. If not provided, it is estimated from a large
sample.}
}
\value{
A `ggplot2` plot showing the standardized sampling
distributions for different sample sizes, compared against
the standard normal curve.
}
\description{
The `demo_clt()` function generates plots to illustrate the
Central Limit Theorem (CLT) using a specified random number
generator. The function displays the sampling distributions
for different sample sizes, standardized and compared against
the standard normal distribution.
}
\examples{
set.seed(123)
demo_clt(runif, n = c(5, 10, 20, 40), nrep = 10000, min = 0,
max = 1)
demo_clt(rgamma, n = c(5, 10, 20, 40), nrep = 10000, shape = 2,
rate = 1, mean = 2, sd = sqrt(2))

}
